add quaternion library from Tobias Simon
See https://github.com/TobiasSimon/quat

Note: Tobias Simon didn't commit this to Space Nerds In Space, nor send a patch for
Space Nerds In Space, this is just his code that I, Stephen M. Cameron, dropped in
here verbatim.  It didn't seem right to have this commit attributed to me as author,
and also not quite right to attribute this commit to Tobias Simon, but the latter
seemed less wrong to me, so that's what I did.

-- steve