gaseous-giganticus: Fix discontinuities in the velocity field.
The problem was in fij_to_xyz().  I had converted i and j
correctly to the correct pair of x, y and z, scaled between 0
and 1, but the remaining coordinate, I scaled between 0 and 500
or so.  So the noise field that was being represented was 6
tiny very separate regions on the face of a sphere (imagine
the central mechanism of a dismantled rubik's cube -- the
six separated square faces on this mechanism were somewhat
like the noise I was sampling.)  This also explained why I
had to pick seemingly excessively large noise scale values.