gaseous-giganticus: optimize curl2 by removing left/right rotation decision
I discovered that due to the value of the "left_right_fudge", the if statement
which relied on this value *always* took one path.  The ostensible purpose of
the if statement was to determine if the noise gradient at a particular point
on the planet was uphill or downhill, then rotate the projected noise vector
left or right depending on that.  However, adjusting the "left_right_fudge"
to 1.0, (which is the more "correct" value) which resulted in roughly half
left rotations and half right rotations (which is what I would expect you'd
want) actually *breaks the program* (divergences in the velocity field).
So, somehow, always rotating left or right seems to be the "correct" answer
(produces seemingly divergence free velocity fields.)  I do not know why this
is the case.  I suspect somehow the math is taking care of this for me already,
and then doing it, in effect, twice, is the wrong thing to do.  But why and
how this might be the case, I do not know.

Nevertheless, I'm committing this mysterous fix.