Address some edge artifacts of flowmap
The flowmap is calculated by getting 2d vectors of
particle movement by remembering a particle's previous
position.  If a particle crosses a cubemap face boundary
the vector calculation is thrown off.  We can remedy some
of this by allowing a particle position calculation to
remain on a preferred face even though it's out of bounds.

This allows velocities to be calculated right up to the
edges of the faces more easily.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>