Use huge pages for large data structures on linux
Use posix_memalign() and madvise() to get "huge pages" for
large data structures (image data, velocity field, and particle
arrays). This can give a slight performance boost due to fewer
TLB misses. There is a --no-huge-pages option to suppress
this optimization.  This optimization is only implemented on
linux.  On other platforms, the --no-huge-pages option has
no effect.

There is now a perfstat script which will run the same workload
with and without huge pages to measure the performance difference.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>