Improve cache-friendliness of initial particle placement
There is a new --cache-aware flag which determines what
fraction of particles should be placed in a cache aware
manner. A value of 1.0 means place 100% of particles in
a cache-aware manner, 0.0 means none, 0.5 means half in
a cache-aware manner, half randomly..

The idea is that particles near each other in the particle
array should also be near each other in the velocity field.

Previously, particle locations were 100% random, which is the
worst possible way to assign particle locations if you care
about the cache friendliness of accesses to the velocity
field arrays.

The reason you may not want all particles placed in a cache
aware manner is that the particles are placed somewhat regularly
and if there are more then 1024 * 1024 * 6 particles, then they
occupy every pixel of the cubemap images.  Jitter is added to
the particle positions to try to alleviate some artifacts
resulting from strictly regular particle placement, but still
some artifacts remain.  Some amount of random particle placement
can alleviate this further.  The default value if none is
specified is 0.7864320000 (1024 * 1024 * 6) / 8000000 -- which
is to say, exactly enough particles to occupy every pixel of
the six cubemap images (~6 million) plus the remainder of
the 8 million placed randomly.

With 100% cache-aware placement of 8M particles, the cache
miss rate is about 52%, while with 100% random placement of
8M particles, the cache miss rate is about 61%.  This can
mean a significant reduction of runtime. When benchmarking
be wary of CPU temperature differences at the start of the
test, lest you be misled by thermal throttling / frequency
scaling, etc.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>