Allow directories in filename prefixes
Previously, when generating images, temporary files were created
with a name beginning with a ".", and then these files were renamed
without the leading "." once all the files were finished.  A problem
with this method is that if the path begins with a "/", adding a "."
in front doesn't do what we want.

Instead, create temporary files with the prefix added without a
leading ".", but insert "-temp-" after the prefix, and then rename
all the files without the inserted "-temp-". This allows the prefix
to contain directories.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>